requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
pandas
numpy
python-dotenv
